Refactored Trade button to be general, added to confirm order screen
