+++ src/screens/orderConfirmation/orderConfirmation.js
@@ -2,6 +2,7 @@ import LoadingSpinner from "../../components/widgets/LoadingSpinner/LoadingSpinn
 import MessageAlert from "../../components/widgets/MessageAlert/MessageAlert";
 import { Container } from "react-bootstrap";
 import { useState } from "react";
+import BottomStickyButton from "../../components/widgets/BottomStickyButton/BottomStickyButton";
 
 function OrderConfirmationPage() {
 
@@ -9,13 +10,17 @@ function OrderConfirmationPage() {
     const [error, setError] = useState("");
 
     return (
-
         <>
             {loading ? <LoadingSpinner /> : error ? <MessageAlert variant='danger'>{error}</MessageAlert> :
-
                 <Container>
                     <h1>Order Confirmation</h1>
+
+
+
+                    <BottomStickyButton text="Confirm Order"></BottomStickyButton>
                 </Container>
+
+
 
             }
         </>
